Drain queue before attempting to get lldp info
TC uses Linux Socket Filtering to filter lldp packets.
LSF work in the way that fiter is applied when frame arrives
on interface - not when it is passed to  userspace with recv().
So it is possible, that if frame did arrive after socket creation
but before filter is applied, such frame will sit in queue, and TC
will attempt to process it (erroneously).
This change makes sure that queue is drained after filter is applied,
prior capturing lldp info.

Change-Id: I60e61748d79aaa188b6872189acc580c539a34ed
Closes-Bug: OPENSTACK-3057